Made config.js config.json and updated default storage mode
Renamed the configuration file to be a JSON rather than a JS so that multi-file editors such as Visual Studio Code and Notepad++ which rely on the file type for syntax highlighting and correction work properly. Also changed the default storage method in the config to be file based storage, as to match the README.md in the repo root defining it as default, and to allow haste-server to work immediately out of the box rather than crashing about memcached not existing, and simply not working once this has been installed but a memcached server is not running.