Upgraded to latest alpine instead of EOL Debian and tested.
Debian Stretch is at the end of life date next year so I upgraded and tested the usability of Alpine. Also verified the default settings of this container still work with less env-vars defined in the Dockerfile.  Next step will be to fully add documentation mentioning the env-vars that are now missing from the official image.